+++ src/components/configEditor/TracesConfig.tsx
@@ -4,7 +4,7 @@ import { ConfigSection, ConfigSubSection } from 'components/experimental/ConfigS
 import { Input, Field } from '@grafana/ui';
 import { OtelVersionSelect } from 'components/queryBuilder/OtelVersionSelect';
 import { ColumnHint, TimeUnit } from 'types/queryBuilder';
-import otel from 'otel';
+import otel, { defaultTraceTable } from 'otel';
 import { LabeledInput } from './LabeledInput';
 import { DurationUnitSelect } from 'components/queryBuilder/DurationUnitSelect';
 import { CHTracesConfig } from 'types/config';
@@ -90,7 +90,7 @@ export const TracesConfig = (props: TraceConfigProps) => {
           onChange={e => onDefaultTableChange(e.currentTarget.value)}
           label={labels.defaultTable.label}
           aria-label={labels.defaultTable.label}
-          placeholder={labels.defaultTable.placeholder}
+          placeholder={defaultTraceTable}
         />
       </Field>
       <ConfigSubSection